+++ armadillo/src/sharedTest/java/at/favre/lib/armadillo/ASecureSharedPreferencesTest.java
@@ -292,7 +292,9 @@ public abstract class ASecureSharedPreferencesTest {
         preferenceSmokeTest(create("recovery", null)
             .recoveryPolicy(true, true).build());
         preferenceSmokeTest(create("recovery", null)
-            .recoveryPolicy(new RecoveryPolicy.Default(true, true)).build());
+            .recoveryPolicy(new SimpleRecoveryPolicy.Default(true, true)).build());
+        preferenceSmokeTest(create("recovery", null)
+            .recoveryPolicy((e, keyHash, base64Encrypted, pwUsed, sharedPreferences) -> System.out.println(e + " " + keyHash + " " + base64Encrypted + " " + pwUsed)).build());
     }
 
     @Test
@@ -359,14 +361,14 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open new shared pref and add some data
         ArmadilloSharedPreferences pref = create(name, clonePassword(currentPassword))
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         pref.edit().putString("k1", "string1").putInt("k2", 2).putStringSet("set", testSet)
-                .putBoolean("k3", true).commit();
+            .putBoolean("k3", true).commit();
         pref.close();
 
         // open again and check if can be used
         pref = create(name, clonePassword(currentPassword))
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
         assertTrue(pref.getBoolean("k3", false));
@@ -375,7 +377,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with old pw and change to new one, all the values should be accessible
         pref = create(name, clonePassword(currentPassword))
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         pref.changePassword(clonePassword(newPassword));
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
@@ -385,7 +387,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with new pw, should be accessible
         pref = create(name, clonePassword(newPassword))
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
         assertTrue(pref.getBoolean("k3", false));
@@ -394,14 +396,14 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with old pw, should throw exception, since cannot decrypt
         pref = create(name, clonePassword(currentPassword))
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         try {
             pref.getString("k1", null);
             fail("should throw exception, since cannot decrypt");
         } catch (SecureSharedPreferenceCryptoException ignored) {
         }
     }
-    
+
     private char[] clonePassword(char[] password) {
         return password == null ? null : password.clone();
     }
@@ -415,14 +417,14 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open new shared pref and add some data
         ArmadilloSharedPreferences pref = create("testChangePassword", "pw1".toCharArray())
-                .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
+            .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
         pref.edit().putString("k1", "string1").putInt("k2", 2).putStringSet("set", testSet)
-                .putBoolean("k3", true).commit();
+            .putBoolean("k3", true).commit();
         pref.close();
 
         // open again and check if can be used
         pref = create("testChangePassword", "pw1".toCharArray())
-                .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
+            .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
         assertTrue(pref.getBoolean("k3", false));
@@ -431,7 +433,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with old pw and old ksFn; change to new one, all the values should be accessible
         pref = create("testChangePassword", "pw1".toCharArray())
-                .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
+            .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
         pref.changePassword("pw2".toCharArray(), new ArmadilloBcryptKeyStretcher(8));
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
@@ -450,7 +452,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with new pw and old ksFn, should throw exception, since cannot decrypt
         pref = create("testChangePassword", "pw2".toCharArray())
-                .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
+            .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
         try {
             pref.getString("k1", null);
             fail("should throw exception, since cannot decrypt");
@@ -459,7 +461,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with old pw and old ksFn, should throw exception, since cannot decrypt
         pref = create("testChangePassword", "pw1".toCharArray())
-                .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
+            .keyStretchingFunction(new BrokenBcryptKeyStretcher(8)).build();
         try {
             pref.getString("k1", null);
             fail("should throw exception, since cannot decrypt");
@@ -471,13 +473,13 @@ public abstract class ASecureSharedPreferencesTest {
     public void testInvalidPasswordShouldNotBeAccessible() {
         // open new shared pref and add some data
         ArmadilloSharedPreferences pref = create("testInvalidPassword", "pw1".toCharArray())
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         pref.edit().putString("k1", "string1").putInt("k2", 2).putBoolean("k3", true).commit();
         pref.close();
 
         // open again and check if can be used
         pref = create("testInvalidPassword", "pw1".toCharArray())
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         assertEquals("string1", pref.getString("k1", null));
         assertEquals(2, pref.getInt("k2", 0));
         assertTrue(pref.getBoolean("k3", false));
@@ -485,7 +487,7 @@ public abstract class ASecureSharedPreferencesTest {
 
         // open with invalid pw, should throw exception, since cannot decrypt
         pref = create("testInvalidPassword", "pw2".toCharArray())
-                .keyStretchingFunction(new FastKeyStretcher()).build();
+            .keyStretchingFunction(new FastKeyStretcher()).build();
         try {
             pref.getString("k1", null);
             fail("should throw exception, since cannot decrypt");